Added animations
I created a few animations for the fishing rod. I created a casting and pulling animation. I acquired a script from a youtuber called Mike's Code, which seems to give a fishing functionality that I'm looking for. I have to scrub through it and fit it to my specific needs though.